feat(i18n): add Korean locale support
Register ko across renderer and main i18n, extend the language picker and
system-locale resolution, bootstrap ko.json from en.json, and generalize
locale catalog tooling to verify every non-English locale file.
