+++ .gitignore
@@ -106,5 +106,6 @@ playwright-report/
 /agent-hooks/
 validation-screenshots/
 
-# Localization bootstrap cache (regenerated by bootstrap:zh-catalog)
+# Localization bootstrap cache (regenerated by bootstrap:*-catalog)
 src/renderer/src/i18n/locales/.zh-catalog-cache.json
+src/renderer/src/i18n/locales/.ko-catalog-cache.json
