+++ src/shared/ui-language.test.ts
@@ -3,6 +3,7 @@ import { describe, expect, it } from 'vitest'
 import {
   UI_LANGUAGE_CHINESE,
   UI_LANGUAGE_ENGLISH,
+  UI_LANGUAGE_KOREAN,
   UI_LANGUAGE_SYSTEM,
   normalizeUiLanguage
 } from './ui-language'
@@ -12,10 +13,11 @@ describe('normalizeUiLanguage', () => {
     expect(normalizeUiLanguage(UI_LANGUAGE_SYSTEM)).toBe('system')
     expect(normalizeUiLanguage(UI_LANGUAGE_ENGLISH)).toBe('en')
     expect(normalizeUiLanguage(UI_LANGUAGE_CHINESE)).toBe('zh')
+    expect(normalizeUiLanguage(UI_LANGUAGE_KOREAN)).toBe('ko')
   })
 
   it('falls back unknown values to system', () => {
-    expect(normalizeUiLanguage('ko')).toBe('system')
+    expect(normalizeUiLanguage('fr')).toBe('system')
     expect(normalizeUiLanguage(null)).toBe('system')
   })
 })
