+++ src/shared/ui-language.ts
@@ -1,16 +1,19 @@
 export const UI_LANGUAGE_SYSTEM = 'system'
 export const UI_LANGUAGE_ENGLISH = 'en'
 export const UI_LANGUAGE_CHINESE = 'zh'
+export const UI_LANGUAGE_KOREAN = 'ko'
 
 export type UiLanguage =
   | typeof UI_LANGUAGE_SYSTEM
   | typeof UI_LANGUAGE_ENGLISH
   | typeof UI_LANGUAGE_CHINESE
+  | typeof UI_LANGUAGE_KOREAN
 
 const UI_LANGUAGE_VALUES = new Set<UiLanguage>([
   UI_LANGUAGE_SYSTEM,
   UI_LANGUAGE_ENGLISH,
-  UI_LANGUAGE_CHINESE
+  UI_LANGUAGE_CHINESE,
+  UI_LANGUAGE_KOREAN
 ])
 
 export function normalizeUiLanguage(value: unknown): UiLanguage {
