+++ src/shared/ui-locale.test.ts
@@ -1,7 +1,12 @@
 import { describe, expect, it } from 'vitest'
 
 import { normalizeSupportedUiLocale, resolveUiLocale, resolveRendererUiLocale } from './ui-locale'
-import { UI_LANGUAGE_CHINESE, UI_LANGUAGE_ENGLISH, UI_LANGUAGE_SYSTEM } from './ui-language'
+import {
+  UI_LANGUAGE_CHINESE,
+  UI_LANGUAGE_ENGLISH,
+  UI_LANGUAGE_KOREAN,
+  UI_LANGUAGE_SYSTEM
+} from './ui-language'
 
 describe('ui-locale', () => {
   it('normalizes supported locale prefixes', () => {
@@ -11,8 +16,12 @@ describe('ui-locale', () => {
     expect(normalizeSupportedUiLocale('zh-SG')).toBe('zh')
   })
 
+  it('normalizes Korean locale prefixes', () => {
+    expect(normalizeSupportedUiLocale('ko-KR')).toBe('ko')
+    expect(normalizeSupportedUiLocale('ko')).toBe('ko')
+  })
+
   it('falls back unsupported locales to English', () => {
-    expect(normalizeSupportedUiLocale('ko-KR')).toBe('en')
     expect(normalizeSupportedUiLocale('fr-FR')).toBe('en')
   })
 
@@ -30,14 +39,20 @@ describe('ui-locale', () => {
     expect(resolveUiLocale(UI_LANGUAGE_CHINESE, 'en-US')).toBe('zh')
   })
 
+  it('resolves explicit Korean independently of system locale', () => {
+    expect(resolveUiLocale(UI_LANGUAGE_KOREAN, 'en-US')).toBe('ko')
+  })
+
   it('maps system locale to the closest supported locale', () => {
     expect(resolveUiLocale(UI_LANGUAGE_SYSTEM, 'en-GB')).toBe('en')
     expect(resolveUiLocale(UI_LANGUAGE_SYSTEM, 'zh-CN')).toBe('zh')
+    expect(resolveUiLocale(UI_LANGUAGE_SYSTEM, 'ko-KR')).toBe('ko')
     expect(resolveUiLocale(UI_LANGUAGE_SYSTEM, 'fr-FR')).toBe('en')
   })
 
   it('uses renderer system locale only for the system setting', () => {
     expect(resolveRendererUiLocale(UI_LANGUAGE_ENGLISH)).toBe('en')
     expect(resolveRendererUiLocale(UI_LANGUAGE_CHINESE)).toBe('zh')
+    expect(resolveRendererUiLocale(UI_LANGUAGE_KOREAN)).toBe('ko')
   })
 })
